# frontend/openapi.json
{
  "openapi": "3.0.1",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [
    { "url": "http://localhost:8000", "description": "Generated server url" }
  ],
...
  "paths": {
...
    "/api/playlists/{id}/tracks": {
      "post": {
        "tags": ["Playlists"],
        "summary": "Add a track to playlist",
        "operationId": "addTrack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddTrackRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PlaylistTrack" }
              }
            }
          }
        }
      }
    },
...
    "/api/playlists/{id}/tracks/{trackId}": {
      "delete": {
        "tags": ["Playlists"],
        "summary": "Remove a track from playlist",
        "operationId": "removeTrack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Playlist": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "userId": { "type": "string", "format": "uuid" },
          "user": { "$ref": "#/components/schemas/User" },
          "isPublic": { "type": "boolean" },
          "shareToken": { "type": "string" },
          "danceStyle": { "type": "string" },
          "subStyle": { "type": "string" },
          "tempoCategory": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "tracks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PlaylistTrack" }
          },
          "collaborators": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PlaylistCollaborator" }
          }
        }
      },
      "PlaylistCollaborator": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "playlistId": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "permission": { "type": "string" },
          "status": { "type": "string" },
          "invitedBy": { "type": "string", "format": "uuid" },
          "invitedAt": { "type": "string", "format": "date-time" },
          "acceptedAt": { "type": "string", "format": "date-time" },
          "playlist": { "$ref": "#/components/schemas/Playlist" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "PlaylistTrack": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "playlistId": { "type": "string", "format": "uuid" },
          "trackId": { "type": "string", "format": "uuid" },
          "position": { "type": "integer", "format": "int32" },
          "addedAt": { "type": "string", "format": "date-time" },
          "playlist": { "$ref": "#/components/schemas/Playlist" },
          "track": { "$ref": "#/components/schemas/Track" }
        }
      },
      "Track": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "isrc": { "type": "string" },
          "durationMs": { "type": "integer", "format": "int32" },
          "tempoBpm": { "type": "number", "format": "float" },
          "createdAt": { "type": "string", "format": "date-time" },
          "hasVocals": { "type": "boolean" },
          "swingRatio": { "type": "number", "format": "float" },
          "articulation": { "type": "number", "format": "float" },
          "bounciness": { "type": "number", "format": "float" },
          "loudness": { "type": "number", "format": "float" },
          "punchiness": { "type": "number", "format": "float" },
          "voiceProbability": { "type": "number", "format": "float" },
          "polskaScore": { "type": "number", "format": "float" },
          "hamboScore": { "type": "number", "format": "float" },
          "bpmStability": { "type": "number", "format": "float" },
          "r1Mean": { "type": "number", "format": "float" },
          "r2Mean": { "type": "number", "format": "float" },
          "r3Mean": { "type": "number", "format": "float" },
          "asymmetryScore": { "type": "number", "format": "float" },
          "asymmetryConsistency": { "type": "number", "format": "float" },
          "patternType": { "type": "string" },
          "ternaryConfidence": { "type": "number", "format": "float" },
          "meterAmbiguous": { "type": "boolean" },
          "isInstrumental": { "type": "boolean" },
          "liltScore": { "type": "number", "format": "float" },
          "liltConsistency": { "type": "number", "format": "float" },
          "liltPattern": {
            "type": "array",
            "items": { "type": "number", "format": "float" }
          },
          "analysisVersion": { "type": "string" },
          "musicGenre": { "type": "string" },
          "genreConfidence": { "type": "number", "format": "float" },
          "isFlagged": { "type": "boolean" },
          "flaggedAt": { "type": "string", "format": "date-time" },
          "flagReason": { "type": "string" },
          "uploaderId": { "type": "string", "format": "uuid" },
          "processingStatus": { "type": "string" },
          "bars": {
            "type": "array",
            "items": { "type": "number", "format": "float" }
          },
          "sections": {
            "type": "array",
            "items": { "type": "number", "format": "float" }
          },
          "sectionLabels": { "type": "array", "items": { "type": "string" } }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "discourseId": { "type": "string" },
          "username": { "type": "string" },
          "displayName": { "type": "string" },
          "avatarUrl": { "type": "string" },
          "role": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "lastLoginAt": { "type": "string", "format": "date-time" },
          "playlists": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Playlist" }
          },
          "collaborations": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PlaylistCollaborator" }
          }
        }
      },
...
      "AddTrackRequest": {
        "type": "object",
        "properties": { "trackId": { "type": "string", "format": "uuid" } }
      },
...
    }
  }
}